+++ src/EightOne.hpp
@@ -54,7 +54,7 @@ struct EightOne : SelectOnClockModule {
 	Trigger _reset;
 	bogaudio::dsp::Timer _timer;
 	int _step;
-	int _select = 0;
+	float _select = 0.0f;
 
 	EightOne() {
 		config(NUM_PARAMS, NUM_INPUTS, NUM_OUTPUTS, NUM_LIGHTS);
